requirements.txt
mysql-connector-python
alive-progress
pandas
flask